Fixed the project sharing functionality

+++ project/src/main/java/edu/tum/ase/project/controller/ProjectController.java
@@ -8,10 +8,15 @@ import org.json.JSONObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.oauth2.client.OAuth2RestOperations;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.client.HttpClientErrorException;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @RestController
 public class ProjectController {
@@ -29,7 +34,9 @@ public class ProjectController {
     @PostMapping("/projects")
     public Project createProject(@RequestParam String name) {
         log.info("createProject('"+name+"')");
-        return projectService.createProject(new Project(name));
+        Authentication authentication = SecurityContextHolder.getContext()
+                .getAuthentication();
+        return projectService.createProject(new Project(name, authentication.getName()));
     }
 
     @GetMapping("/projects/{name}")
@@ -39,7 +46,7 @@ public class ProjectController {
 
     @GetMapping("/projects")
     public List<Project> getAllProjects(){
-        return projectService.getProjects();
+        return projectService.getProjects().stream().filter(p -> p.getUserIds().contains(SecurityContextHolder.getContext().getAuthentication().getName())).collect(Collectors.toList());
     }
 
     @DeleteMapping("/{projectName}")
@@ -47,48 +54,52 @@ public class ProjectController {
         // when a project gets deleted, all of its source files should also be deleted
         // this surely can be done more elegant using the correct @OnDelete annotations in the entity models
         // I was to stupid for that though.. feel free to improve this :)
-        sourceFileService.getAllFilesOfProject(projectName)
-                .stream()
-                .forEach(sourceFile -> sourceFileService.removeSourceFile(sourceFile));
-        projectService.deleteProject(projectService.findByName(projectName));
+        Authentication authentication = SecurityContextHolder.getContext()
+                .getAuthentication();
+        if (projectService.findByName(projectName).getUserIds().contains(authentication.getName())) {
+            sourceFileService.getAllFilesOfProject(projectName)
+                    .stream()
+                    .forEach(sourceFile -> sourceFileService.removeSourceFile(sourceFile));
+            projectService.deleteProject(projectService.findByName(projectName));
+        } else {
+            throw new HttpClientErrorException(HttpStatus.FORBIDDEN);
+        }
     }
 
     @PutMapping("/projects/{oldName}")
     public Project renameProject(@PathVariable String oldName, @RequestBody String newName) {
-        Project newProject = new Project(newName);
-        projectService.createProject(newProject);
-        sourceFileService.getAllFilesOfProject(oldName)
-                .stream()
-                .forEach(sourceFile ->
-                    sourceFileService.createSourceFile(new SourceFile(newProject,
-                            sourceFile.getName(), sourceFile.getCode()))
-                );
-        deleteProject(oldName);
-        return newProject;
+        if (!oldName.equals(newName)) {
+            Project newProject = new Project(newName, SecurityContextHolder.getContext().getAuthentication().getName());
+            projectService.createProject(newProject);
+            sourceFileService.getAllFilesOfProject(oldName)
+                    .stream()
+                    .forEach(sourceFile ->
+                            sourceFileService.createSourceFile(new SourceFile(newProject,
+                                    sourceFile.getName(), sourceFile.getCode()))
+                    );
+            deleteProject(oldName);
+            return newProject;
+        } else {
+            throw new HttpClientErrorException(HttpStatus.NOT_ACCEPTABLE);
+        }
     }
 
-
-    @GetMapping("/projects/{projectName}")
-    public void foo(@PathVariable String projectName, @PathVariable String userId)
-    {
+    @GetMapping("/projects/{projectName}/{userId}")
+    public void shareProject(@PathVariable String projectName, @PathVariable String userId) {
         String jsonStr = operations.getForObject(
-                "https://gitlab.example.com/api/v4/search?scope=users&search="+ userId,
+                "https://gitlab.lrz.de/api/v4/search?scope=users&search="+ userId,
                 String.class
         );
 
-
-        JSONObject data = new JSONObject(jsonStr);
+        assert jsonStr != null;
+        JSONObject data = new JSONObject(jsonStr.substring(1, jsonStr.length() - 1));
         if(data.has("username") && data.get("username").equals(userId))
         {
                 projectService.addUserId(projectName, userId);
         }
         else
         {
-            System.out.println(jsonStr);
+            throw new HttpClientErrorException(HttpStatus.NOT_FOUND);
         }
     }
-
-
-
-
 }
